+++ .changeset/2ba86628/changes.md
@@ -1,1 +1,1 @@
-- Adding smart cards to the renderer and on document load
+- ED-6698: Adding smart cards to the renderer and on document load